% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{'Brute-force' creation of multipanel plots}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{One or more plot objects to be paneled}

\item{plotlist}{A list structure of plot objects}

\item{cols}{The number of columns to create in the multipanel}

\item{layout}{A matrix object to force a layout (optional)}
}
\value{
Prints the plots in ... as a multipanel plot
}
\description{
'Brute-force' creation of multipanel plots
}
\details{
Originally found at:
    http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
with no copyright information.

ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
- cols:   Number of columns in layout
- layout: A matrix specifying the layout. If present, 'cols' is ignored.

If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}

